wrapper: remove fortify for cuda compilation
When using clang to compile CUDA (>= 9.2) using foritfy results in
errors. This change adds the wrapper_unsupported_hardening_flags array
to the wrapper, which is used to disable hardening flags from withinthe
wrapper. This list is then used to disable fortify when compiling for
CUDA using '-x cuda'.
NOTE: '-x cuda' can be omitted if the file end in '.cu', but I don't
know a good way to check for this in the wrapper, so this is not
checked.